Refactor split out yaep_print.c
