+++ scripts/build_xmq_from_parts.sh
@@ -82,6 +82,8 @@ do_part yaep_structs
 do_part yaep_util
 do_part yaep_symbols
 do_part yaep_terminal_bitset
+do_part yaep_print
+do_part yaep_tree
 
 
 do_version
