+++ src/main/c/parts/yaep_util.c
@@ -31,6 +31,7 @@
 #include "yaep_symbols.h"
 #include "yaep_terminal_bitset.h"
 #include "yaep_tree.h"
+#include "yaep_print.h"
 
 #endif
 
@@ -115,421 +116,6 @@ int find_matched_length(YaepParseState *ps,
     }
 
     return matched_length;
-}
-
-void print_core(MemBuffer*mb, YaepStateSetCore *c, YaepStateSet *s)
-{
-    membuffer_printf(mb, "core %d %s {", c->id, c->term->hr);
-
-    for (int i = 0; i < c->num_dotted_rules; ++i)
-    {
-        // num_started_dotted_rules;
-        YaepDottedRule *dotted_rule = c->dotted_rules[i];
-        membuffer_printf(mb, " %d[%d]",
-                         dotted_rule->id,
-                         dotted_rule->dot_j);
-
-        if (i < c->num_all_matched_lengths)
-        {
-            int len = -1;
-            if (s) len = s->matched_lengths[i];
-            membuffer_printf(mb, "(len%d,pid%d)",
-                             len,
-                             c->parent_dotted_rule_ids[i]);
-        }
-    }
-    membuffer_printf(mb, "}");
-}
-
-void print_coresymbvects(MemBuffer*mb, YaepParseState *ps, YaepCoreSymbToPredComps *v)
-{
-    membuffer_printf(mb, "coresymbvect %d %s preds: ", v->core->id, v->symb->hr);
-    for (int i = 0; i < v->predictions.len; ++i)
-    {
-        int dotted_rule_id = v->predictions.ids[i];
-        YaepDottedRule *dotted_rule = v->core->dotted_rules[dotted_rule_id];
-        membuffer_printf(mb, " (%d)%s", dotted_rule_id, dotted_rule->rule->lhs->hr);
-    }
-    membuffer_printf(mb, " comps:");
-    for (int i = 0; i < v->completions.len; ++i)
-    {
-        int dotted_rule_id = v->completions.ids[i];
-        YaepDottedRule *dotted_rule = v->core->dotted_rules[dotted_rule_id];
-        membuffer_printf(mb, " (%d)%s", dotted_rule_id, dotted_rule->rule->lhs->hr);
-    }
-}
-
-/* The following function prints RULE with its translation(if TRANS_P) to file F.*/
-void rule_print(MemBuffer *mb, YaepParseState *ps, YaepRule *rule, bool trans_p)
-{
-    int i;
-
-    if (rule->mark != 0
-        && rule->mark != ' '
-        && rule->mark != '-'
-        && rule->mark != '@'
-        && rule->mark != '^'
-        && rule->mark != '*')
-    {
-        membuffer_append(mb, "\n(yaep) internal error bad rule: ");
-        symbol_print(mb, rule->lhs, false);
-        debug_mb("ixml=", mb);
-        free_membuffer_and_free_content(mb);
-        assert(false);
-    }
-
-    char m = rule->mark;
-    if (m >= 32 && m < 127)
-    {
-        membuffer_append_char(mb, m);
-    }
-    symbol_print(mb, rule->lhs, false);
-    if (strcmp(rule->lhs->repr, rule->anode))
-    {
-        membuffer_append_char(mb, '(');
-        membuffer_append(mb, rule->anode);
-        membuffer_append_char(mb, ')');
-    }
-    membuffer_append(mb, " → ");
-    int cost = rule->anode_cost;
-    while (cost > 0)
-    {
-        membuffer_append_char(mb, '<');
-        cost--;
-    }
-    for (i = 0; i < rule->rhs_len; i++)
-    {
-        char m = rule->marks[i];
-        if (m >= 32 && m < 127)
-        {
-            membuffer_append_char(mb, m);
-        }
-        else
-        {
-            if (!m) membuffer_append(mb, "  ");
-            else    assert(false);
-        }
-        symbol_print(mb, rule->rhs[i], false);
-    }
-    /*
-      if (false && trans_p)
-      {
-      fprintf(f, "      | ");
-      if (rule->anode != NULL)
-      {
-      fprintf(f, "%s(", rule->anode);
-      }
-      for(i = 0; i < rule->trans_len; i++)
-      {
-      for(j = 0; j < rule->rhs_len; j++)
-      {
-      if (rule->order[j] == i)
-      {
-      fprintf(f, " %d:", j);
-      // TODO symbol_print(f, rule->rhs[j], false);
-      break;
-      }
-      }
-      if (j >= rule->rhs_len)
-      {
-      fprintf(f, " nil");
-      }
-      }
-      if (rule->anode != NULL)
-      {
-      fprintf(f, " )");
-      }
-    */
-}
-
-/* The following function prints RULE to file F with dot in position pos.
-   Pos == 0 means that the dot is all the way to the left in the starting position.
-   Pos == rhs_len means that the whole rule has been matched. */
-void print_rule_with_dot(MemBuffer *mb, YaepParseState *ps, YaepRule *rule, int pos)
-{
-    int i;
-
-    assert(pos >= 0 && pos <= rule->rhs_len);
-
-    symbol_print(mb, rule->lhs, false);
-    membuffer_append(mb, " → ");
-    for(i = 0; i < rule->rhs_len; i++)
-    {
-        membuffer_append(mb, i == pos ? " · " : " ");
-        symbol_print(mb, rule->rhs[i], false);
-    }
-}
-
-void print_rule(MemBuffer *mb, YaepParseState *ps, YaepRule *rule)
-{
-    int i;
-
-    symbol_print(mb, rule->lhs, false);
-    membuffer_append(mb, " → ");
-    for(i = 0; i < rule->rhs_len; i++)
-    {
-        membuffer_append_char(mb, ' ');
-        symbol_print(mb, rule->rhs[i], false);
-    }
-}
-
-/* The following function prints the dotted_rule.
-   Print the lookahead set if lookahead_p is true. */
-void print_dotted_rule(MemBuffer *mb,
-                              YaepParseState *ps,
-                              int from_i,
-                              YaepDottedRule *dotted_rule,
-                              int matched_length,
-                              int parent_id,
-                              const char *why)
-{
-    char buf[256];
-
-    snprintf(buf, 256, "(s%d,d%d) ", from_i, dotted_rule->id);
-    membuffer_append(mb, buf);
-    print_rule_with_dot(mb, ps, dotted_rule->rule, dotted_rule->dot_j);
-
-    if (matched_length == 0)
-    {
-        if (dotted_rule->rule->rhs_len == dotted_rule->dot_j)
-        {
-            if (dotted_rule->rule->rhs_len == 0) membuffer_append(mb, " ε");
-            snprintf(buf, 256, " complete[%d-%d/%d]", ps->tok_i, 1+ps->tok_i, ps->input_len);
-            membuffer_append(mb, buf);
-        }
-        else
-        {
-            snprintf(buf, 256, " prediction[%d-%d/%d]", ps->tok_i, 1+ps->tok_i, ps->input_len);
-            membuffer_append(mb, buf);
-        }
-    }
-    else if (matched_length > 0)
-    {
-        if (dotted_rule->rule->rhs_len == dotted_rule->dot_j)
-        {
-            snprintf(buf, 256, " complete[%d-%d/%d]", 1+ps->tok_i-matched_length, 1+ps->tok_i, ps->input_len);
-            membuffer_append(mb, buf);
-        }
-        else
-        {
-            snprintf(buf, 256, " partial[%d-%d/%d]", 1+ps->tok_i-matched_length, 1+ps->tok_i, ps->input_len);
-            membuffer_append(mb, buf);
-        }
-    }
-    else
-    {
-        membuffer_append(mb, " n/a[]");
-    }
-    int cost = dotted_rule->rule->anode_cost;
-    if (cost > 0)
-    {
-        while (cost-- > 0) membuffer_append(mb, "<");
-        membuffer_append(mb, " ");
-    }
-    if (dotted_rule->rule->lhs->empty_p ||
-        dotted_rule->empty_tail_p ||
-        dotted_rule->info)
-    {
-        membuffer_append(mb, " ");
-    }
-
-    membuffer_printf(mb, "{%s:%s", dotted_rule->info, why);
-    if (dotted_rule->rule->lhs->empty_p || dotted_rule->empty_tail_p)
-    {
-        membuffer_append(mb, " ");
-    }
-    if (dotted_rule->empty_tail_p)
-    {
-        membuffer_append(mb, " empty_tail");
-    }
-    if (dotted_rule->rule->lhs->empty_p)
-    {
-        membuffer_append(mb, " empty_rule");
-    }
-
-    if (parent_id >= 0) membuffer_printf(mb, " parent=d%d", parent_id);
-    membuffer_printf(mb, " ml=%d", matched_length);
-    membuffer_append(mb, "}");
-
-    if (*why)
-    {
-        if (ps->run.grammar->lookahead_level != 0 && matched_length >= 0)
-        {
-            membuffer_append(mb, "    ");
-            terminal_bitset_print(mb, ps, dotted_rule->lookahead);
-        }
-    }
-}
-
-/* The following recursive function prints NODE into file F and prints
-   all its children(if debug_level < 0 output format is for graphviz).*/
-void print_yaep_node(YaepParseState *ps, FILE *f, YaepTreeNode *node)
-{
-    YaepTreeNodeVisit*trans_visit_node;
-    YaepTreeNode*child;
-    int i;
-
-    assert(node != NULL);
-    trans_visit_node = visit_node(ps, node);
-    if (trans_visit_node->num >= 0)
-    {
-        return;
-    }
-    trans_visit_node->num = -trans_visit_node->num - 1;
-    if (ps->run.debug) fprintf(f, "%7d: ", trans_visit_node->num);
-    switch(node->type)
-    {
-    case YAEP_NIL:
-        if (ps->run.debug)
-            fprintf(f, "EMPTY\n");
-        break;
-    case YAEP_ERROR:
-        if (ps->run.debug > 0)
-            fprintf(f, "ERROR\n");
-        break;
-    case YAEP_TERM:
-        if (ps->run.debug)
-            fprintf(f, "TERMINAL: code=%d, repr=%s, mark=%d %c\n", node->val.terminal.code,
-                    symb_find_by_code(ps, node->val.terminal.code)->repr, node->val.terminal.mark, node->val.terminal.mark>32?node->val.terminal.mark:' ');
-        break;
-    case YAEP_ANODE:
-        if (ps->run.debug)
-        {
-            fprintf(f, "ABSTRACT: %c%s(", node->val.anode.mark?node->val.anode.mark:' ', node->val.anode.name);
-            for(i = 0;(child = node->val.anode.children[i]) != NULL; i++)
-            {
-                fprintf(f, " %d", canon_node_id(visit_node(ps, child)->num));
-            }
-            fprintf(f, ")\n");
-        }
-        else
-        {
-            for(i = 0;(child = node->val.anode.children[i]) != NULL; i++)
-            {
-                fprintf(f, "  \"%d: %s\" -> \"%d: ", trans_visit_node->num,
-                         node->val.anode.name,
-                        canon_node_id(visit_node(ps, child)->num));
-                switch(child->type)
-                {
-                case YAEP_NIL:
-                    fprintf(f, "EMPTY");
-                    break;
-                case YAEP_ERROR:
-                    fprintf(f, "ERROR");
-                    break;
-                case YAEP_TERM:
-                    fprintf(f, "%s",
-                            symb_find_by_code(ps, child->val.terminal.code)->repr);
-                    break;
-                case YAEP_ANODE:
-                    fprintf(f, "%s", child->val.anode.name);
-                    break;
-                case YAEP_ALT:
-                    fprintf(f, "ALT");
-                    break;
-                default:
-                    assert(false);
-                }
-                fprintf(f, "\";\n");
-            }
-        }
-        for (i = 0;(child = node->val.anode.children[i]) != NULL; i++)
-        {
-            print_yaep_node(ps, f, child);
-        }
-        break;
-    case YAEP_ALT:
-        if (ps->run.debug)
-        {
-            fprintf(f, "ALTERNATIVE: node=%d, next=",
-                    canon_node_id(visit_node(ps, node->val.alt.node)->num));
-            if (node->val.alt.next != NULL)
-                fprintf(f, "%d\n",
-                        canon_node_id(visit_node(ps, node->val.alt.next)->num));
-            else
-                fprintf(f, "nil\n");
-        }
-        else
-        {
-            fprintf(f, "  \"%d: ALT\" -> \"%d: ", trans_visit_node->num,
-                    canon_node_id(visit_node(ps, node->val.alt.node)->num));
-            switch(node->val.alt.node->type)
-            {
-            case YAEP_NIL:
-                fprintf(f, "EMPTY");
-                break;
-            case YAEP_ERROR:
-                fprintf(f, "ERROR");
-                break;
-            case YAEP_TERM:
-                fprintf(f, "%s",
-                        symb_find_by_code(ps, node->val.alt.node->val.terminal.code)->
-                         repr);
-                break;
-            case YAEP_ANODE:
-                fprintf(f, "%s", node->val.alt.node->val.anode.name);
-                break;
-            case YAEP_ALT:
-                fprintf(f, "ALT");
-                break;
-            default:
-                assert(false);
-            }
-            fprintf(f, "\";\n");
-            if (node->val.alt.next != NULL)
-            {
-                fprintf(f, "  \"%d: ALT\" -> \"%d: ALT\";\n",
-                        trans_visit_node->num,
-                        canon_node_id(visit_node(ps, node->val.alt.next)->num));
-            }
-        }
-        print_yaep_node(ps, f, node->val.alt.node);
-        if (node->val.alt.next != NULL)
-        {
-            print_yaep_node(ps, f, node->val.alt.next);
-        }
-        break;
-    default:
-        assert(false);
-    }
-}
-
-/* The following function prints parse tree with ROOT.*/
-void print_parse(YaepParseState *ps, FILE* f, YaepTreeNode*root)
-{
-    ps->map_node_to_visit = create_hash_table(ps->run.grammar->alloc,
-                                              ps->input_len* 2,
-                                              trans_visit_node_hash,
-                                              trans_visit_node_eq);
-
-    ps->num_nodes_visits = 0;
-    OS_CREATE(ps->node_visits_os, ps->run.grammar->alloc, 0);
-    print_yaep_node(ps, f, root);
-    OS_DELETE(ps->node_visits_os);
-    delete_hash_table(ps->map_node_to_visit);
-}
-
-/* The following function prints SET to file F.  If NOT_YET_STARTED_P is true
-   then print all dotted_rules.  The dotted_rules are printed with the
-   lookahead set if LOOKAHEAD_P.  SET_DIST is used to print absolute
-   matched_lengths of not-yet-started dotted_rules. */
-void print_state_set(MemBuffer *mb,
-                     YaepParseState *ps,
-                     YaepStateSet *state_set,
-                     int from_i)
-{
-    StateVars vars;
-    fetch_state_vars(ps, state_set, &vars);
-
-    membuffer_printf(mb, "state=%d core=%d", vars.state_id, vars.core_id);
-
-    int dotted_rule_id = 0;
-    for(; dotted_rule_id < vars.num_dotted_rules; dotted_rule_id++)
-    {
-        int matched_length = find_matched_length(ps, state_set, &vars, dotted_rule_id);
-        membuffer_append(mb, "\n");
-        print_dotted_rule(mb, ps, from_i, vars.dotted_rules[dotted_rule_id], matched_length, -1, "woot3");
-    }
 }
 
 bool is_not_rule(YaepSymbol *symb)
