+++ src/main/c/xmq.c
@@ -53,6 +53,7 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #include"parts/yaep_symbols.h"
 #include"parts/yaep_terminal_bitset.h"
 #include"parts/yaep_tree.h"
+#include"parts/yaep_print.h"
 
 // XMQ STRUCTURES ////////////////////////////////////////////////
 
@@ -4968,6 +4969,7 @@ char *xmqLinePrintf(XMQLineConfig *lc, const char *element_name, ...)
 #include"parts/yaep_symbols.c"
 #include"parts/yaep_terminal_bitset.c"
 #include"parts/yaep_tree.c"
+#include"parts/yaep_print.c"
 #include"parts/yaep.c"
 
 #endif
